+++ app/Http/Controllers/BookController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Book;
+use App\Http\Controllers\API\APIBaseController as APIBaseController;
 use Illuminate\Http\Request;
+use Validator;
 
-class BookController extends Controller
+class BookController extends APIBaseController
 {
     /**
      * Display a listing of the resource.
@@ -14,7 +16,11 @@ class BookController extends Controller
      */
     public function index()
     {
-        return Book::all();
+        $books = Book::paginate(20);
+        if (count($books) < 1) {
+            return $this->sendMessage('Found 0 book');
+        }
+        return $this->sendData($books->toArray());
     }
 
     /**
@@ -31,14 +37,46 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
-        $book = Book::create($request->all());
+        $db_book = Book::get();
+        foreach ($db_book as $result) {
+            if ($request->slug == $result->slug) {
+                return $this->sendError('This book already exits, please enter another name & slug !');
+            }
+        }
+        $input = $request->all();
+        $validator = Validator::make($input, [
+            'name' => 'required',
+            'slug' => 'required',
+            'image' => 'required',
+            'price' => 'required',
+            'highlights' => 'required',
+            'description' => 'required',
+            'quantity' => 'required',
+            'id_author' => 'required',
+            'id_category' => 'required',
+        ], [
+            'name.required' => 'Please enter book name',
+            'slug.required' => 'Please enter slug',
+            'image.required' => 'Please enter book image',
+            'price.required' => 'Please enter price',
+            'highlights.required' => 'Please choose type highlights',
+            'description.required' => 'Please enter book description',
+            'quantity.required' => 'Please enter quantity this book',
+            'id_author.required' => 'Please choose author',
+            'id_category.required' => 'Please choose category',
+        ]);
+        if ($validator->fails()) {
+            return $this->sendError('Validation Error.', $validator->errors());
+        }
+        $book = Book::create($input);
         if ($request->hasFile('image')) {
             $file = $request->file('image');
             $file->move('./images', $file->getClientOriginalName('image'));
             $image = $file->getClientOriginalName('image');
-            $book->update(['image' => $image]);
+            $book->image = $image;
         }
-        return response()->json($book, 201);
+        $book->save();
+        return $this->sendResponse($book->toArray(), 'Book created successfully.');
     }
 
     /**
@@ -49,7 +87,11 @@ class BookController extends Controller
      */
     public function show($id)
     {
-        return Book::find($id);
+        $book = Book::find($id);
+        if (is_null($book)) {
+            return $this->sendError('Book not found.');
+        }
+        return $this->sendData($book->toArray());
     }
 
     /**
@@ -67,15 +109,58 @@ class BookController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $book = Book::findOrFail($id);
-        $book->update($request->all());
+        $book = Book::find($id);
+        if (is_null($book)) {
+            return $this->sendError('Book not found.');
+        }
+        if ($book->slug !== $request->slug) {
+            $db_book = Book::get();
+            foreach ($db_book as $result) {
+                if ($request->slug == $result->slug) {
+                    return $this->sendError('This book already exits, please enter another name & slug !');
+                }
+            }
+        }
+        $input = $request->all();
+        $validator = Validator::make($input, [
+            'name' => 'required',
+            'slug' => 'required',
+            'price' => 'required',
+            'highlights' => 'required',
+            'description' => 'required',
+            'quantity' => 'required',
+            'id_author' => 'required',
+            'id_category' => 'required',
+        ], [
+            'name.required' => 'Please enter book name',
+            'slug.required' => 'Please enter slug',
+            'price.required' => 'Please enter price',
+            'highlights.required' => 'Please choose type highlights',
+            'description.required' => 'Please enter book description',
+            'quantity.required' => 'Please enter quantity this book',
+            'id_author.required' => 'Please choose author',
+            'id_category.required' => 'Please choose category',
+        ]);
+        if ($validator->fails()) {
+            return $this->sendError('Validation Error.', $validator->errors());
+        }
+
         if ($request->hasFile('image')) {
             $file = $request->file('image');
             $file->move('./images', $file->getClientOriginalName('image'));
             $image = $file->getClientOriginalName('image');
-            $book->update(['image' => $image]);
+            $book->image = $image;
         }
-        return response()->json($book, 200);
+        $book->name = $input['name'];
+        $book->slug = $input['slug'];
+        $book->price = $input['price'];
+        $book->highlights = $input['highlights'];
+        $book->description = $input['description'];
+        $book->quantity = $input['quantity'];
+        $book->id_author = $input['id_author'];
+        $book->id_category = $input['id_category'];
+        $book->save();
+        return $this->sendResponse($book->toArray(), 'Book updated successfully');
     }
 
     /**
@@ -86,11 +171,12 @@ class BookController extends Controller
      */
     public function destroy($id)
     {
-        $book = Book::findOrFail($id);
+        $book = Book::find($id);
+        if (is_null($book)) {
+            return $this->sendError('Book not found.');
+        }
         $book->delete();
-
-        return response()->json(null, 204);
+        return $this->sendResponse($id, 'Book deleted successfully');
     }
 
-    
 }
