+++ app/Http/Controllers/TagController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\API\APIBaseController as APIBaseController;
 use App\Tag;
 use Illuminate\Http\Request;
+use Validator;
 
-class TagController extends Controller
+class TagController extends APIBaseController
 {
     /**
      * Display a listing of the resource.
@@ -14,7 +16,11 @@ class TagController extends Controller
      */
     public function index()
     {
-        return Tag::all();
+        $tags = Tag::paginate(15);
+        if (count($tags) < 1) {
+            return $this->sendMessage('Found 0 tags');
+        }
+        return $this->sendData($tags->toArray());
     }
 
     /**
@@ -31,12 +37,31 @@ class TagController extends Controller
      */
     public function store(Request $request)
     {
+        $db_tags = Tag::get();
+        foreach ($db_tags as $result) {
+            if ($result->slug == $request->slug) {
+                return $this->sendError('This tag already exits !');
+            }
+        }
+        $input = $request->all();
+        $validator = Validator::make($input, [
+            'name' => 'required',
+            'slug' => 'required',
+            'description' => 'required',
+        ], [
+            'name.required' => 'Please enter name',
+            'slug.required' => 'Please enter slug',
+            'description.required' => 'Please enter description',
+        ]);
+        if ($validator->fails()) {
+            return $this->sendError('Validation Error.', $validator->errors());
+        }
         $tag = new Tag;
-        $tag->name = $request->name;
-        $tag->slug = $request->slug;
-        $tag->description = $request->description;
+        $tag->name = $input['name'];
+        $tag->slug = $input['slug'];
+        $tag->description = $input['description'];
         $tag->save();
-        return response()->json($tag, 201);
+        return $this->sendResponse($tag->toArray(), 'Tag created successfully');
     }
 
     /**
@@ -47,7 +72,11 @@ class TagController extends Controller
      */
     public function show($id)
     {
-        return Tag::find($id);
+        $tag = Tag::find($id);
+        if (is_null($tag)) {
+            return $this->sendError('Tag not found.');
+        }
+        return $this->sendData($tag->toArray());
     }
 
     /**
@@ -66,9 +95,34 @@ class TagController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $tag = Tag::findOrFail($id);
-        $tag->update($request->all());
-        return response()->json($tag, 200);
+        $tag = Tag::find($id);
+        if (is_null($tag)) {
+            return $this->sendError('Tag not found.');
+        }
+        if ($tag->slug !== $request->slug) {
+            $db_tags = Tag::get();
+            foreach ($db_tags as $result) {
+                if ($result->slug == $request->slug) {
+                    return $this->sendError('This tag already exits !');
+                }
+            }
+        }
+        $input = $request->all();
+        $validator = Validator::make($input, [
+            'name' => 'required',
+            'slug' => 'required',
+            'description' => 'required',
+        ], [
+            'name.required' => 'Please enter name',
+            'slug.required' => 'Please enter slug',
+            'description.required' => 'Please enter description',
+        ]);
+        if ($validator->fails()) {
+            return $this->sendError('Validation Error.', $validator->errors());
+        }
+
+        $tag->update($input);
+        return $this->sendResponse($tag->toArray(), 'Tag updated successfully');
     }
 
     /**
@@ -79,9 +133,11 @@ class TagController extends Controller
      */
     public function destroy($id)
     {
-        $tag = Tag::findOrFail($id);
+        $tag = Tag::find($id);
+        if (is_null($tag)) {
+            return $this->sendError('Tag not found.');
+        }
         $tag->delete();
-
-        return response()->json(null, 204);
+        return $this->sendResponse($id, 'Tag deleted successfully');
     }
 }
